Added Dijkstra's
Added Dijkstra's pathfinding algorithm.
The visualisation for this algorithm is very similar to that of the dfs algorithm but this uses a priority queue and can take into account different weight of edges. However all the edges between nodes in this program have the same weight so there is little difference visually