Small feature
Added another maze type, "Imperfect".
This is my first attempt at a proper maze generating algorithm and is imperfect because it leaves a small few areas of the maze inacessible.

Also changed the sizing of the window and changed the options for number of rows. This was so that there is always an odd number of rows/columns.